requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0

# RL and Robotics
gymnasium>=0.28.0
mujoco>=3.0.0
stable-baselines3>=2.0.0
torch>=2.0.0

# Utilities
pyyaml>=6.0
opencv-python>=4.5.0
pillow>=9.0.0
tqdm>=4.62.0

# Logging and monitoring
tensorboard>=2.10.0
wandb>=0.13.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.20.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
